+++ HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ---------------
 
+2.2.0 (?)
++++++++++
+
+**Bugfixes**
+
+- Fixed bug in ``ArchiveSession`` object where domains weren't getting set properly for cookies.
+  This caused archive.org cookies to be sent to other domains.
+
 2.1.0 (2021-08-25)
 ++++++++++++++++++